+++ Sources/StreamVideo/WebRTC/PeerConnection/Extensions/RTCIceConnectionState+CustomStringConvertible.swift
@@ -1,0 +1,32 @@
+//
+// Copyright © 2024 Stream.io Inc. All rights reserved.
+//
+
+import Foundation
+import StreamWebRTC
+
+extension RTCIceConnectionState: CustomStringConvertible {
+
+    public var description: String {
+        switch self {
+        case .new:
+            return "new"
+        case .checking:
+            return "checking"
+        case .connected:
+            return "connected"
+        case .completed:
+            return "completed"
+        case .failed:
+            return "failed"
+        case .disconnected:
+            return "disconnected"
+        case .closed:
+            return "closed"
+        case .count:
+            return "count"
+        @unknown default:
+            return "unknown"
+        }
+    }
+}
