+++ Sources/StreamVideo/WebRTC/PeerConnection/Extensions/RTCIceGatheringState+CustomStringConvertible.swift
@@ -1,0 +1,21 @@
+//
+// Copyright © 2024 Stream.io Inc. All rights reserved.
+//
+
+import Foundation
+import StreamWebRTC
+
+extension RTCIceGatheringState: CustomStringConvertible {
+    public var description: String {
+        switch self {
+        case .new:
+            return "new"
+        case .gathering:
+            return "gathering"
+        case .complete:
+            return "complete"
+        @unknown default:
+            return "unknown/default"
+        }
+    }
+}
