+++ Sources/StreamVideo/WebRTC/PeerConnection/Models/PeerConnectionType.swift
@@ -1,0 +1,10 @@
+//
+// Copyright © 2024 Stream.io Inc. All rights reserved.
+//
+
+import Foundation
+
+enum PeerConnectionType: String {
+    case subscriber
+    case publisher
+}
